refactor(frontend): move Home tile sub-components into local components/ folders
