+++ autogpt_platform/frontend/src/app/(platform)/home/components/NeedsYou/components/AttentionRow.tsx
