+++ autogpt_platform/frontend/src/app/(platform)/home/components/NowNext/NowNext.tsx
@@ -1,14 +1,10 @@
-import { Calendar03Icon, Clock01Icon } from "@hugeicons/core-free-icons";
-import Link from "next/link";
-import type { HomeActiveTask } from "@/app/api/__generated__/models/homeActiveTask";
+import { Calendar03Icon } from "@hugeicons/core-free-icons";
 import type { HomeDashboardResponse } from "@/app/api/__generated__/models/homeDashboardResponse";
-import type { HomeUpcomingTask } from "@/app/api/__generated__/models/homeUpcomingTask";
 import { Icon } from "@/components/atoms/Icon/Icon";
 import { Text } from "@/components/atoms/Text/Text";
-import { ExpertAvatar } from "@/components/molecules/ExpertAvatar/ExpertAvatar";
-import { cn } from "@/lib/utils";
 import { HomeTile } from "../HomeTile/HomeTile";
-import { formatRunningFor, formatUntil } from "./helpers";
+import { ActiveRow } from "./components/ActiveRow";
+import { UpcomingRow } from "./components/UpcomingRow";
 
 interface Props {
   dashboard: HomeDashboardResponse;
@@ -71,76 +67,5 @@ export function NowNext({ dashboard, className }: Props) {
         )}
       </div>
     </HomeTile>
-  );
-}
-
-function ActiveRow({ item }: { item: HomeActiveTask }) {
-  const content = (
-    <>
-      <span className="absolute left-0 top-1/2 z-10 flex size-8 -translate-y-1/2 items-center justify-center rounded-full bg-white">
-        {item.expert ? (
-          <ExpertAvatar
-            name={item.expert.name}
-            avatarUrl={item.expert.avatar_url}
-            size={32}
-          />
-        ) : (
-          <span className="size-2.5 rounded-full bg-primary" />
-        )}
-      </span>
-      <div className="min-w-0 flex-1">
-        <Text variant="body-medium" className="truncate text-zinc-900">
-          {item.title}
-        </Text>
-        <Text variant="small" className="truncate text-zinc-500">
-          {item.status === "queued"
-            ? "Queued"
-            : (formatRunningFor(item.started_at) ?? "Running now")}
-        </Text>
-      </div>
-    </>
-  );
-  const classes = "relative flex items-center gap-3 rounded-xl py-3 pl-12 pr-2";
-
-  if (!item.link) return <div className={classes}>{content}</div>;
-  return (
-    <Link
-      href={item.link}
-      className={cn(
-        classes,
-        "transition-colors hover:bg-zinc-50 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-zinc-400",
-      )}
-    >
-      {content}
-    </Link>
-  );
-}
-
-function UpcomingRow({ item }: { item: HomeUpcomingTask }) {
-  return (
-    <div className="relative flex items-center gap-3 rounded-xl py-3 pl-12 pr-2">
-      <span className="absolute left-0 top-1/2 z-10 flex size-8 -translate-y-1/2 items-center justify-center rounded-full bg-zinc-100 text-zinc-500">
-        <Icon
-          icon={item.kind === "followup" ? Calendar03Icon : Clock01Icon}
-          size={15}
-          aria-hidden="true"
-        />
-      </span>
-      <div className="min-w-0 flex-1">
-        <Text variant="body-medium" className="truncate text-zinc-900">
-          {item.title}
-        </Text>
-        <Text variant="small" className="truncate text-zinc-500">
-          {item.expert?.name ??
-            (item.kind === "followup" ? "Follow-up" : "Scheduled task")}
-        </Text>
-      </div>
-      <Text
-        variant="small"
-        className="shrink-0 font-medium tabular-nums text-zinc-700"
-      >
-        {formatUntil(item.next_run_time)}
-      </Text>
-    </div>
   );
 }
